+++ src/realm/object-store/sync/app.cpp
@@ -917,7 +917,7 @@ void App::handle_possible_redirect_response(Request&& request, const Response& r
                                             UniqueFunction<void(const Response&)>&& completion)
 {
     // If the response contains a redirection, then process it
-    if (util::HTTPStatus(response.http_status_code) == util::HTTPStatus::MovedPermanently) {
+    if (sync::HTTPStatus(response.http_status_code) == sync::HTTPStatus::MovedPermanently) {
         handle_redirect_response(std::move(request), response, std::move(completion));
     }
     else {