+++ src/realm/sync/network/network.cpp
@@ -69,7 +69,7 @@
 #endif
 
 using namespace realm::util;
-using namespace realm::util::network;
+using namespace realm::sync::network;
 
 
 namespace {
@@ -2507,9 +2507,7 @@ bool ReadAheadBuffer::read(char*& begin, char* end, int delim, std::error_code& 
 }
 
 
-namespace realm {
-namespace util {
-namespace network {
+namespace realm::sync::network {
 
 std::string host_name()
 {
@@ -2609,6 +2607,4 @@ std::string ResolveErrorCategory::message(int value) const
     return {};
 }
 
-} // namespace network
-} // namespace util
-} // namespace realm
+} // namespace realm::sync::network